+++ playground/rewrite-unoptimized/test-package/es/nested/foo.js
